+++ backend/handlers/list.go
@@ -1,9 +1,11 @@
 package handlers
 
 import (
+	"fmt"
 	"net/http"
 
 	"github.com/labstack/echo/v4"
+	"github.com/stashsphere/backend/handlers/params"
 	"github.com/stashsphere/backend/middleware"
 	"github.com/stashsphere/backend/resources"
 	"github.com/stashsphere/backend/services"
@@ -81,10 +83,40 @@ func (lh *ListHandler) ListHandlerShow(c echo.Context) error {
 }
 
 type ListsParams struct {
-	Page           uint64   `query:"page"`
-	PerPage        uint64   `query:"perPage"`
-	FilterOwnerIds []string `query:"filterOwnerId"`
-	Paginate       *bool    `query:"paginate"`
+	Page           uint64         `query:"page"`
+	PerPage        uint64         `query:"perPage"`
+	FilterOwnerIds []string       `query:"filterOwnerId"`
+	Paginate       *bool          `query:"paginate"`
+	Order          []params.Order `query:"order" validate:"max=2,dive"`
+}
+
+func ParamsOrderToListServiceOrder(order []params.Order) ([]services.ListOrder, error) {
+	res := make([]services.ListOrder, len(order))
+	for i, o := range order {
+		so := services.ListOrder{}
+		switch o.FieldName {
+		case "access_reason":
+			so.Field = services.OrderFieldAccessReason
+			break
+		case "created_at":
+			so.Field = services.OrderFieldCreatedAt
+			break
+		default:
+			return nil, utils.ParameterError{Err: fmt.Errorf("Invalid field.")}
+		}
+		switch o.FieldSortBy {
+		case "desc":
+			so.Direction = services.OrderDescending
+			break
+		case "asc":
+			so.Direction = services.OrderAscending
+			break
+		default:
+			return nil, utils.ParameterError{Err: fmt.Errorf("Invalid field.")}
+		}
+		res[i] = so
+	}
+	return res, nil
 }
 
 func (lh *ListHandler) ListHandlerIndex(c echo.Context) error {
@@ -99,12 +131,20 @@ func (lh *ListHandler) ListHandlerIndex(c echo.Context) error {
 	if err := c.Bind(&params); err != nil {
 		return &utils.ParameterError{Err: err}
 	}
+	if err := c.Validate(&params); err != nil {
+		return &utils.ParameterError{Err: err}
+	}
 	if params.PerPage == 0 {
 		params.PerPage = 50
 	}
 	paginate := true
 	if params.Paginate != nil && *params.Paginate == true {
 		paginate = true
+	}
+
+	order, err := ParamsOrderToListServiceOrder(params.Order)
+	if err != nil {
+		return err
 	}
 	result, err := lh.listService.GetListsForUser(c.Request().Context(),
 		services.GetListsForUserParams{
@@ -113,6 +153,7 @@ func (lh *ListHandler) ListHandlerIndex(c echo.Context) error {
 			Page:           params.Page,
 			FilterOwnerIds: params.FilterOwnerIds,
 			Paginate:       paginate,
+			Order:          order,
 		},
 	)
 	if err != nil {